requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
geoalchemy2==0.14.2
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0.post1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
PyJWT==2.8.0
python-dotenv==1.0.0

# Dependencias para scraper
shapely==2.0.2
